fix tests

+++ iterators/range.js
@@ -133,7 +133,9 @@ module.exports = class RangeIterator {
       if (!isKey) {
         if (!top.node.children.length) continue
         const node = await top.node.getChildNode(n)
-        top.node.children[n].value = null // unlink it to save memory
+        if (top.node.block.seq < this.batch.core.length) {
+          top.node.children[n].value = null // unlink it to save memory
+        }
         this.stack.push({ i: this._reverse ? node.keys.length << 1 : 0, node })
         continue
       }
